# openapi.yaml
openapi: 3.0.3
info:
  title: City Pulse API
  version: 0.1.0
  description: >
    REST API for City Pulse — users (city location = San Diego only), regions,
    events, trends (most interacted events by region), and interactions (likes, comments, attending).
    Authenticate via POST /api/auth/login and use the returned access_token as Bearer token for GET /api/auth/me.
servers:
  - url: http://localhost:8000
    description: Local development
paths:
...
  /api/regions/{region_id_or_name}/events:
    get:
      summary: List events in a region
      tags: [Regions]
      parameters:
        - name: region_id_or_name
          in: path
          required: true
          description: Region ID (0) or name ('san diego', 'san-diego')
          schema:
            oneOf:
              - type: integer
                example: 0
              - type: string
                example: "san diego"
      responses:
        "200":
          description: Events in the region
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/EventRead"
        "400":
          description: Unknown region
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDetail"
  /api/regions/{region_id_or_name}/users:
    get:
      summary: List users in a region
      tags: [Regions]
      parameters:
        - name: region_id_or_name
          in: path
          required: true
          description: Region ID (0) or name ('san diego', 'san-diego')
          schema:
            oneOf:
              - type: integer
              - type: string
      responses:
        "200":
          description: Users in the region
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserRead"
        "400":
          description: Unknown region
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDetail"

  # ---------------------------------------------------------------------------
  # Events
  # ---------------------------------------------------------------------------
...
  /api/interactions:
    get:
      summary: List events in a region with interaction counts and comments
      tags: [Interactions]
      parameters:
        - name: region
          in: query
          schema:
            type: string
            default: "san diego"
          description: 'san diego or 0'
        - name: skip
          in: query
          schema:
            type: integer
            default: 0
            minimum: 0
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
            minimum: 1
            maximum: 100
      responses:
        "200":
          description: Events with likes_count, comments_count, attendance_count, comments[]
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/EventWithInteractionsRead"
        "400":
          description: Unknown region
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDetail"
  /api/interactions/events/{event_id}/likes:
    put:
      summary: Add a like from a user to an event (idempotent)
      tags: [Interactions]
      parameters:
        - name: event_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InteractionLikeBody"
      responses:
        "200":
          description: Success (or already liked)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "404":
          description: Event not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDetail"
        "422":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationError"
    delete:
      summary: Remove a user's like from an event
      tags: [Interactions]
      parameters:
        - name: event_id
          in: path
          required: true
          schema:
            type: integer
        - name: user_id
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Like removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "404":
          description: Like not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDetail"
  /api/interactions/events/{event_id}/comments:
    put:
      summary: Add a comment from a user to an event
      tags: [Interactions]
      parameters:
        - name: event_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InteractionCommentBody"
      responses:
        "201":
          description: Created comment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommentRead"
        "404":
          description: Event not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDetail"
        "422":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationError"
  /api/interactions/events/{event_id}/comments/{comment_id}:
    delete:
      summary: Remove a user's comment (user must own the comment)
      tags: [Interactions]
      parameters:
        - name: event_id
          in: path
          required: true
          schema:
            type: integer
        - name: comment_id
          in: path
          required: true
          schema:
            type: integer
        - name: user_id
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Comment removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "403":
          description: Cannot delete another user's comment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDetail"
        "404":
          description: Comment not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDetail"
  /api/interactions/events/{event_id}/attending:
    put:
      summary: Mark user as attending an event (idempotent)
      tags: [Interactions]
      parameters:
        - name: event_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InteractionAttendingBody"
      responses:
        "200":
          description: Success (or already attending)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "404":
          description: Event not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDetail"
        "422":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationError"
    delete:
      summary: Remove a user's attendance from an event
      tags: [Interactions]
      parameters:
        - name: event_id
          in: path
          required: true
          schema:
            type: integer
        - name: user_id
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Attendance removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "404":
          description: Attendance record not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDetail"
components:
  schemas:
    ErrorDetail:
      type: object
      description: Standard error body
      properties:
        detail:
          oneOf:
            - type: string
            - type: array
              items:
                type: object
                properties:
                  loc:
                    type: array
                    items: {}
                  msg:
                    type: string
                  type:
                    type: string
    ValidationError:
      type: object
      properties:
        detail:
          type: array
          items:
            type: object
            properties:
              loc:
                type: array
              msg:
                type: string
              type:
                type: string
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
...
    UserRead:
      type: object
      description: User profile (API response)
      properties:
        id:
          type: integer
          example: 0
        name:
          type: string
          example: "John"
        email:
          type: string
          format: email
          example: "john@example.com"
        city_location:
          type: string
          nullable: true
          example: "san diego"
        created_at:
          type: string
          format: date-time
...
    EventRead:
      type: object
      description: Event in a region
      properties:
        id:
          type: integer
        region_id:
          type: integer
        user_id:
          type: integer
          nullable: true
        title:
          type: string
        content:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
...
    EventWithInteractionsRead:
      type: object
      description: Event with interaction counts and comments list
      properties:
        id:
          type: integer
        region_id:
          type: integer
        user_id:
          type: integer
          nullable: true
        title:
          type: string
        content:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        likes_count:
          type: integer
          default: 0
        comments_count:
          type: integer
          default: 0
        attendance_count:
          type: integer
          default: 0
        comments:
          type: array
          items:
            $ref: "#/components/schemas/CommentRead"
    CommentRead:
      type: object
      description: Single comment on an event
      properties:
        id:
          type: integer
        user_id:
          type: integer
        event_id:
          type: integer
        text:
          type: string
        created_at:
          type: string
          format: date-time
    InteractionLikeBody:
      type: object
      required: [user_id]
      properties:
        user_id:
          type: integer
          description: User who is liking the event
    InteractionCommentBody:
      type: object
      required: [user_id, text]
      properties:
        user_id:
          type: integer
          description: User who is commenting
        text:
          type: string
          minLength: 1
          maxLength: 5000
    InteractionAttendingBody:
      type: object
      required: [user_id]
      properties:
        user_id:
          type: integer
          description: User who is attending
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Use access_token from POST /api/auth/login